+++ packages/moonstone/VirtualList/VirtualListBaseNative.js
@@ -263,7 +263,6 @@ class VirtualListCoreNative extends Component {
 	cc = []
 	scrollPosition = 0
 	isScrolledBy5way = false
-	isItemReadyForRestore = true
 
 	containerClass = null
 	contentRef = null
@@ -438,7 +437,6 @@ class VirtualListCoreNative extends Component {
 
 		this.maxFirstIndex = dataSize - numOfItems;
 		this.curDataSize = dataSize;
-		this.isItemReadyForRestore = (preservedIndex <= dataSize);
 
 		// reset children
 		this.cc = [];
@@ -446,11 +444,11 @@ class VirtualListCoreNative extends Component {
 		this.updateMoreInfo(dataSize, scrollPosition);
 
 		if (this.restoreLastFocused &&
-			numOfItems > 0 &&
-			(this.preservedIndex < moreInfo.firstVisibleIndex || this.preservedIndex > moreInfo.lastVisibleIndex) && this.isItemReadyForRestore) {
+			numOfItems > 0 && (preservedIndex < dataSize) &&
+			(preservedIndex < moreInfo.firstVisibleIndex || preservedIndex > moreInfo.lastVisibleIndex)) {
 			// If we need to restore last focus and the index is beyond the screen,
 			// we call `scrollTo` to create DOM for it.
-			this.props.cbScrollTo({index: this.preservedIndex, animate: false});
+			this.props.cbScrollTo({index: preservedIndex, animate: false});
 		} else {
 			newFirstIndex = this.calculateFirstIndex(props, wasFirstIndexMax, dataSizeDiff);
 		}